added option for pick if you want to format the EFI_partition or not

+++ install.sh
@@ -112,11 +112,16 @@ partitioning() {
     echo "EFI partition (/dev/sda1): "
     read EFI_PARTITION
 
+    echo "you want to format the EFI partition? (y/n)"
+    read format_efi_answer
+
     echo "ROOT partition (/dev/sda2): "
     read ROOT_PARTITION
 
-    echo "formating $EFI_PARTITION as FAT32"
-    mkfs.fat -F32 $EFI_PARTITION
+    if [[format_efi_answer == "y" || format_efi_answer == "Y"]]; then
+        echo "formating $EFI_PARTITION as FAT32"
+        mkfs.fat -F32 $EFI_PARTITION
+    fi
 
     echo "formating $ROOT_PARTITION as BTRFS"
     mkfs.btrfs -L arch $ROOT_PARTITION
